LayoutTests/fast/selectors/tagname-adjacent-backtracking-case-sensitivity-html-expected.txt

Test the correctness of the backtracking optimizations with mixed case tag names in the selector.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelectorAll("SIBLING.target + SIBLING + sibling ~ target").length is 1
PASS successfullyParsed is true

TEST COMPLETE
